Changing notifications usability
