+++ mobile/screens/PrivateNotifications.js
@@ -1,7 +1,7 @@
-import { TextField } from 'react-native-material-textfield';;
+
+import { TextField } from 'react-native-material-textfield';
 import { Constants, ImagePicker, Permissions } from 'expo';
-import { NOTIFICATIONS_API } from '../const/Const'
-import React, { Component, Children } from 'react';
+import React, { Component } from 'react';
 import {
   ActivityIndicator,
   View,
@@ -12,23 +12,24 @@ import {
   Alert,
   Image,
   KeyboardAvoidingView,
-  Picker,
+  Picker
 } from 'react-native';
+import { NOTIFICATIONS_API } from './tab_navigator/car_defense/screens/TabNavigator/const/Const'
 import { Icon } from "native-base";
 
 
-const tk = 0;
-
-export default class PublicNotifications extends Component {
+export default class PrivateNotifications extends Component {
 
   constructor(props) {
     super(props);
 
+    this.plateRef = this.updateRef.bind(this, 'plate');
     this.messageRef = this.updateRef.bind(this, 'message');
     this.onFocus = this.onFocus.bind(this);
 
     this.state = {
       title: '',
+      plate: '',
       message: '',
       image: null,
       uploading: false,
@@ -37,7 +38,6 @@ export default class PublicNotifications extends Component {
   updateTitle = (title) => {
     this.setState({ title: title})
     }
-    
   onFocus() {
     let { errors = {} } = this.state;
 
@@ -53,6 +53,13 @@ export default class PublicNotifications extends Component {
   }
 
   onChangeText(words) {
+    ['plate']
+      .map((text) => ({ text, ref: this[text] }))
+      .forEach(({ text, ref }) => {
+        if (ref && ref.isFocused()) {
+          this.setState({ [text]: words });
+        }
+      });
     ['message']
       .map((text) => ({ text, ref: this[text] }))
       .forEach(({ text, ref }) => {
@@ -68,8 +75,24 @@ export default class PublicNotifications extends Component {
 
   onPressButton = () => {
     let errors = {};
-    const url = NOTIFICATIONS_API + '/send_emergency_push_message/' //function send_emergency_push_message url
-    let i = 0;
+    const url = NOTIFICATIONS_API + '/send_push_message/' //function send_push_message url
+    let errorPlate = false;
+    let errorMessage = false;
+
+    ['plate']
+      .forEach((text) => {
+        let value = this[text].value();
+
+        if (!value) {
+          errors[text] = 'Campo obrigatório.';
+          errorPlate = true;
+        } else {
+          if ('plate' === text && value.length < 8) {
+            errors[text] = 'Placa inválida. Ex.:AAA-0000';
+            errorPlate = true;
+          }
+        }
+      });
 
     ['message']
       .forEach((text) => {
@@ -77,19 +100,20 @@ export default class PublicNotifications extends Component {
 
         if (!value) {
           errors[text] = 'Campo obrigatório.';
-          i += 1;
+          errorMessage = true;
         } else {
           if ('message' === text && value.length < 5) {
             errors[text] = 'Forneça mais detalhes do ocorrido.';
-            i += 1;
+            errorMessage = true;
           }
         }
       });
 
-    if (i == 0) {
+    if (errorPlate == false && errorMessage == false) {
       let notification = JSON.stringify({
-        sender_id: tk,
+        //sender_id: tk,
         title: this.state.title,
+        plate: this.state.plate,
         message: this.state.message,
         image: this.state.image
       })
@@ -106,13 +130,13 @@ export default class PublicNotifications extends Component {
       }).then(response => { return response.json() }
       ).then(jsonResponse => {
         console.log(jsonResponse);
+        Alert.alert("Notificação enviada!")
       }
       ).catch(error => {
         console.log(error)
+        Alert.alert("Placa não existe!")
       })
-      Alert.alert("Alerta enviado!")
     }
-
     this.setState({ errors });
   }
 
@@ -123,6 +147,7 @@ export default class PublicNotifications extends Component {
   render() {
     let { image } = this.state;
     let { errors = {}, ...data } = this.state;
+    let { plate = 'text' } = data;
     let { message = 'text' } = data;
 
     return (
@@ -133,17 +158,45 @@ export default class PublicNotifications extends Component {
             <View style={styles.header}>
               <View style={styles.headerContent}>
                 <Image style={styles.avatar}
-                  source={require('../../../../../../images/alert.png')}
+                  source={require('../images/notification.png')}
                 />
-                <Text style={styles.name}>Envie aqui os alertas</Text>
+                <Text style={styles.name}>Envie aqui as notificões</Text>
               </View>
-            </View>
-            <Picker selectedValue= {this.state.title} onValueChange={this.updateTitle} style={{ color: "white" }} mode="dropdown">           
-              <Picker.Item label="Alerta Geral" value="Alerta Geral" />
-              <Picker.Item label="Roubo" value="Roubo" />
-              <Picker.Item label="Incêndio" value="Incendio" />
-              <Picker.Item label ="Tempestade" value="Tempestade" />
-            </Picker>           
+            </View>    
+            <Picker
+             selectedValue= {this.state.title} 
+             onValueChange={this.updateTitle} 
+             style={{ color: "white" }} 
+             mode="dropdown"
+            >        
+              <Picker.Item label="Notificação" value="Notificacao"/>
+              <Picker.Item label="Vidro aberto" value="Vidro aberto" />
+              <Picker.Item label="Farol aceso" value="Farol aceso" />
+              <Picker.Item label ="Pneu furado" value="Pneu furado" />
+              <Picker.Item label ="Alarme disparado" value="Alarme disparado" />
+              <Picker.Item label ="Local estacionado" value="Local estacionado" />
+              
+
+            </Picker>
+            <TextField
+              ref={this.plateRef}
+              value={data.plate}
+              autoCorrect={false}
+              enablesReturnKeyAutomatically={true}
+              onFocus={this.onFocus}
+              onChangeText={(plate) => this.setState({ plate })}
+              // onChangeText={this.onChangeText}
+              // onSubmitEditing={this.onSubmitPLate}
+              returnKeyType='next'
+              label='Placa'
+              tintColor="white"
+              underlineColorAndroid="transparent"
+              maxLength={8}
+              autoCapitalize="characters"
+              error={errors.plate}
+              textColor="white"
+              placeholderTextColor="white"
+            />
             <TextField
               ref={this.messageRef}
               value={data.message}
@@ -161,6 +214,7 @@ export default class PublicNotifications extends Component {
               labelPadding={10}
               placeholderTextColor="white"
             />
+            {/* <Text style={styles.text}>Adicionar imagem</Text> */}
             <View style={styles.alternativeLayoutButtonContainer}>
               <TouchableOpacity
                 color="#8bd4da"
@@ -194,7 +248,7 @@ export default class PublicNotifications extends Component {
                 onPress={this.onPressButton}
                 containerViewStyle={{ width: '40%' }}
               >
-                <Text style={{ color: '#8bd4da', fontWeight: '800' }} >Enviar</Text>
+                <Text style={{ color: '#62969A', fontWeight: '800' }} >Enviar</Text>
               </TouchableOpacity>
             </View>
           </View>
@@ -209,7 +263,7 @@ export default class PublicNotifications extends Component {
       return (
         <View
           style={styles.maybeRenderUploading}>
-          <ActivityIndicator color="#540b71" size="large" />
+          <ActivityIndicator color="#313869" size="large" />
         </View>
       );
     }
@@ -251,7 +305,7 @@ export default class PublicNotifications extends Component {
     alert('Copied image URL to clipboard');
   };
 
-  _takePhoto = async () => {
+  _takePhoto = async () => { //Done
     const {
       status: cameraPerm
     } = await Permissions.askAsync(Permissions.CAMERA);
@@ -368,7 +422,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-evenly',
     paddingTop: 4,
-    paddingRight: 1,
+    paddingRight: 4,
     marginTop: 20
   },
   button: {
@@ -401,8 +455,7 @@ const styles = StyleSheet.create({
   container1: {
     paddingTop: 20,
     flexDirection: 'row',
-    justifyContent: 'center',
-    marginBottom: 40
+    justifyContent: 'center'
   },
   header: {
     marginTop: 25,
@@ -448,7 +501,7 @@ const styles = StyleSheet.create({
   },
   name: {
     fontSize: 18,
-    color: "white",
+    color: "#8bd4da",
     fontWeight: '800',
   },
   headerContent: {
