--- mobile/screens/tab_navigator/car_defense/screens/TabNavigator/PrivateNotifications/PrivateNotificationsNavigator.js
@@ -1,24 +1,0 @@
-import React, { Component } from "react";
-import {StackNavigator} from 'react-navigation'
-
-import PrivateNotifications from './PrivateNotifications'
-
-
-export default class PrivateNotificationsNavigator extends Component {
-
-    render() {
-        return (
-            <PrivateNotificationsStackNavigator />
-        );
-    }
-}
-
-const PrivateNotificationsStackNavigator = new StackNavigator({
-
-    PrivateNotifications: {
-        screen: PrivateNotifications,
-        navigationOptions: ({ navigation }) => ({
-            header: null,
-        }),
-    },
-})
